add mappers + change dto fields where needed

+++ src/main/java/at/qe/skeleton/dtos/LimitDTO.java
@@ -13,6 +13,6 @@ public record LimitDTO(
 
         float humMin,
 
-        int co2Max
+        float co2Max
 ) {
 }
